+++ b2sdk/transfer/emerge/write_intent.py
@@ -67,7 +67,7 @@ class WriteIntent:
         """
         return self.outbound_source.is_upload()
 
-    def get_large_file_sha1(self) -> Optional[Sha1HexDigest]:
+    def get_content_sha1(self) -> Optional[Sha1HexDigest]:
         """
         Return a 40-character string containing the hex SHA1 checksum, which can be used as the `large_file_sha1` entry.
 
@@ -78,7 +78,7 @@ class WriteIntent:
 
         :rtype str:
         """
-        return self.outbound_source.get_large_file_sha1()
+        return self.outbound_source.get_content_sha1()
 
     @classmethod
     def wrap_sources_iterator(cls, outbound_sources_iterator):