+++ b2sdk/utils/__init__.py
@@ -16,17 +16,18 @@ import re
 import shutil
 import tempfile
 import time
-import concurrent.futures as futures
-from abc import abstractmethod
-from collections.abc import Iterator
 from decimal import Decimal
 from itertools import chain
-from typing import Tuple, Optional
+from typing import Any, Iterator, List, NewType, Optional, Tuple, TypeVar
 from urllib.parse import quote, unquote_plus
 
 from logfury.v1 import DefaultTraceAbstractMeta, DefaultTraceMeta, limit_trace_arguments, disable_trace, trace_call
 
-Sha1HexDigest = str
+Sha1HexDigest = NewType('Sha1HexDigest', str)
+T = TypeVar('T')
+# TODO: When we drop Python 3.7 support, this should be replaced
+#   with typing.Protocol that exposes read method.
+ReadOnlyStream = Any
 
 
 def b2_url_encode(s):
@@ -90,15 +91,12 @@ def choose_part_ranges(content_length, minimum_part_size):
     return parts
 
 
-# TODO: When dropping support for Python 3.7 use typing.Protocol to specify what
-# methods input_stream, digest and returned objects should expose.  The defined
-# protocols could be used as type hints in this and related functions.
-def update_digest_from_stream(digest: object, input_stream: object, content_length: int) -> object:
+def update_digest_from_stream(digest: T, input_stream: ReadOnlyStream, content_length: int) -> T:
     """
     Update and return `digest` with data read from `input_stream`
 
-    :param digest: a digest object, which exposes an `update()` method
-    :param input_stream: stream object, which exposes a `read()` method
+    :param digest: a digest object, which exposes an `update(bytes)` method
+    :param input_stream: stream object, which exposes a `read(int|None)` method
     :param content_length: expected length of the stream
     :type content_length: int
     """
@@ -116,49 +114,56 @@ def update_digest_from_stream(digest: object, input_stream: object, content_leng
     return digest
 
 
-def hex_sha1_of_stream(input_stream: object, content_length: int) -> Sha1HexDigest:
+def hex_sha1_of_stream(input_stream: ReadOnlyStream, content_length: int) -> Sha1HexDigest:
     """
     Return the 40-character hex SHA1 checksum of the first content_length
     bytes in the input stream.
 
-    :param input_stream: stream object, which exposes read() method
+    :param input_stream: stream object, which exposes read(int|None) method
     :param content_length: expected length of the stream
     :type content_length: int
     :rtype: str
     """
-    return update_digest_from_stream(hashlib.sha1(), input_stream, content_length).hexdigest()
+    return Sha1HexDigest(
+        update_digest_from_stream(
+            hashlib.sha1(),
+            input_stream,
+            content_length,
+        ).hexdigest()
+    )
 
 
 def hex_sha1_of_unlimited_stream(
-    input_stream: object, limit: Optional[int] = None
-) -> Sha1HexDigest:
+    input_stream: ReadOnlyStream,
+    limit: Optional[int] = None,
+) -> Tuple[Sha1HexDigest, int]:
     block_size = 1024 * 1024
-    content_length = 0
+    offset = 0
     digest = hashlib.sha1()
     while True:
         if limit is not None:
-            to_read = min(limit - content_length, block_size)
+            to_read = min(limit - offset, block_size)
         else:
             to_read = block_size
         data = input_stream.read(to_read)
         data_len = len(data)
         if data_len > 0:
             digest.update(data)
-            content_length += data_len
+            offset += data_len
         if data_len < to_read:
-            return digest.hexdigest(), content_length
+            return Sha1HexDigest(digest.hexdigest()), offset
 
 
 def hex_sha1_of_file(path_) -> Sha1HexDigest:
     with open(path_, 'rb') as file:
-        return hex_sha1_of_unlimited_stream(file)
+        return hex_sha1_of_unlimited_stream(file)[0]
 
 
 def hex_sha1_of_bytes(data: bytes) -> Sha1HexDigest:
     """
     Return the 40-character hex SHA1 checksum of the data.
     """
-    return hashlib.sha1(data).hexdigest()
+    return Sha1HexDigest(hashlib.sha1(data).hexdigest())
 
 
 def hex_md5_of_bytes(data: bytes) -> str:
@@ -455,9 +460,9 @@ def current_time_millis():
     return int(round(time.time() * 1000))
 
 
-def iterator_peek(iterator: Iterator, count: int) -> Tuple[list, Iterator]:
+def iterator_peek(iterator: Iterator[T], count: int) -> Tuple[List[T], Iterator[T]]:
     """
-    Get the `count` first elements yielded by `iterator`.
+    Get up to the `count` first elements yielded by `iterator`.
 
     The function will read `count` elements from `iterator` or less if the end is reached first.  Returns a tuple
     consisting of a list of retrieved elements and an iterator equivalent to the input iterator.