+++ CHANGELOG.md
@@ -9,8 +9,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 * Authorizing a key for a single bucket ensures that this bucket is cached
 * `Bucket.ls` operation supports wildcard matching strings
+* Documentation for AbstractUploadSource and its children
 * Add `large_file_sha1` support
 * Add support for incremental upload and sync
+
+### Fixed
+* Removed information about replication being in closed beta
 
 ### Infrastructure
 * Additional tests for listing files/versions